Conversion to HTML/CSS
In addition to converting this version also introduces The strudel syntax generation tool Which includes individual note sequences as well as chord sequences with associated formatting as well as quantized duration syntax. It also introduces a simple keyboard layout that spans three octaves. (for users without a midi input device).

The Keyboard changes color control based on the tonic note and scale that has been selected by the user.